move RocksDBIndexingEnabler|Disabler into own file
