+++ arangod/RocksDBEngine/RocksDBMethods.h
@@ -25,8 +25,6 @@
 
 #include "RocksDBEngine/RocksDBCommon.h"
 
-#include <memory>
-
 namespace rocksdb {
 class Slice;
 }  // namespace rocksdb
@@ -91,67 +89,6 @@ class RocksDBMethods {
                                        RocksDBKey const&) = 0;
 
   virtual void PutLogData(rocksdb::Slice const&) = 0;
-};
-
-// INDEXING MAY ONLY BE DISABLED IN TOPLEVEL AQL TRANSACTIONS
-// THIS IS BECAUSE THESE TRANSACTIONS WILL EITHER READ FROM
-// OR (XOR) WRITE TO A COLLECTION. IF THIS PRECONDITION IS
-// VIOLATED THE DISABLED INDEXING WILL BREAK GET OPERATIONS.
-struct IndexingDisabler {
-  // will only be active if condition is true
-
-  IndexingDisabler() = delete;
-  IndexingDisabler(IndexingDisabler&&) = delete;
-  IndexingDisabler(IndexingDisabler const&) = delete;
-  IndexingDisabler& operator=(IndexingDisabler const&) = delete;
-  IndexingDisabler& operator=(IndexingDisabler&&) = delete;
-
-  IndexingDisabler(RocksDBMethods* meth, bool condition) : _meth(nullptr) {
-    if (condition) {
-      bool disabledHere = meth->DisableIndexing();
-      if (disabledHere) {
-        _meth = meth;
-      }
-    }
-  }
-
-  ~IndexingDisabler() {
-    if (_meth) {
-      _meth->EnableIndexing();
-    }
-  }
-
- private:
-  RocksDBMethods* _meth;
-};
-
-// if only single indices should be enabled during operations
-struct IndexingEnabler {
-  // will only be active if condition is true
-
-  IndexingEnabler() = delete;
-  IndexingEnabler(IndexingEnabler&&) = delete;
-  IndexingEnabler(IndexingEnabler const&) = delete;
-  IndexingEnabler& operator=(IndexingEnabler const&) = delete;
-  IndexingEnabler& operator=(IndexingEnabler&&) = delete;
-
-  IndexingEnabler(RocksDBMethods* meth, bool condition) : _meth(nullptr) {
-    if (condition) {
-      bool enableHere = meth->EnableIndexing();
-      if (enableHere) {
-        _meth = meth;
-      }
-    }
-  }
-
-  ~IndexingEnabler() {
-    if (_meth) {
-      _meth->DisableIndexing();
-    }
-  }
-
- private:
-  RocksDBMethods* _meth;
 };
 
 }  // namespace arangodb
